move memoize to root utils
